+++ test/NonceManager.t.sol
@@ -13,7 +13,7 @@ contract NonceManagerTest is DelegationHandler {
     function test_getSeq_succeeds() public view {
         // Start with nonce 0, which has key = 0 and sequence = 0
         uint256 nonce = 0;
-        uint192 nonceKey = uint192(nonce >> 64); // Extract key (high 192 bits)
+        uint256 nonceKey = uint192(nonce >> 64); // Extract key (high 192 bits)
         uint256 expectedSeq = uint256(uint64(nonce));
         assertEq(signerAccount.getSeq(nonceKey), expectedSeq);
     }
@@ -41,7 +41,7 @@ contract NonceManagerTest is DelegationHandler {
     }
 
     function test_invalidateNonce_revertsWithExcessiveInvalidation() public {
-        uint192 nonceKey = 0;
+        uint256 nonceKey = 0;
         uint64 sequence = uint64(type(uint16).max) + 1; // Use a high sequence number
         uint256 nonce = (uint256(nonceKey) << 64) | sequence;
 
@@ -51,7 +51,7 @@ contract NonceManagerTest is DelegationHandler {
     }
 
     function test_invalidateNonce_succeeds() public {
-        uint192 nonceKey = 0;
+        uint256 nonceKey = 0;
         uint64 sequence = type(uint16).max;
         uint256 nonce = (uint256(nonceKey) << 64) | sequence;
 
@@ -71,7 +71,7 @@ contract NonceManagerTest is DelegationHandler {
         assertEq(signerAccount.getSeq(nonceKey), sequence);
     }
 
-    function test_fuzz_invalidateNonce(uint192 nonceKey, uint16 sequence) public {
+    function test_fuzz_invalidateNonce(uint256 nonceKey, uint16 sequence) public {
         // Skip sequences that would overflow when incremented
         sequence = uint16(bound(sequence, 1, type(uint16).max));
 
@@ -88,7 +88,7 @@ contract NonceManagerTest is DelegationHandler {
     /// forge-config: default.isolate = true
     /// forge-config: ci.isolate = true
     function test_invalidateNonce_gas() public {
-        uint192 nonceKey = 0;
+        uint256 nonceKey = 0;
         uint64 sequence = type(uint16).max;
         uint256 nonce = (uint256(nonceKey) << 64) | sequence;
